+++ app/src/main/java/com/example/breakingbadwiki/fragment/OthersFragment.kt
@@ -1,12 +1,16 @@
 package com.example.breakingbadwiki.fragment
 
+import android.app.AlertDialog
 import android.content.Intent
 import android.os.Bundle
+import android.util.Log
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.Toast
+import androidx.core.view.isVisible
 import androidx.fragment.app.Fragment
+import androidx.fragment.app.FragmentTransaction
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
 import cn.pedant.SweetAlert.SweetAlertDialog
@@ -15,7 +19,7 @@ import com.example.breakingbadwiki.activity.MainActivity2
 import com.example.breakingbadwiki.adapter.ExploreAdapter
 import com.example.breakingbadwiki.adapter.ItemEvents
 import com.example.breakingbadwiki.data.ItemPost
-import com.example.breakingbadwiki.databinding.FragmentGroupsBinding
+import com.example.breakingbadwiki.databinding.DialogAddItemBinding
 import com.example.breakingbadwiki.databinding.FragmentOthersBinding
 
 class OthersFragment : Fragment(), ItemEvents {
@@ -68,6 +72,52 @@ class OthersFragment : Fragment(), ItemEvents {
             LinearLayoutManager(context, RecyclerView.VERTICAL, false)
         binding.recyclerOthers.adapter = myAdapter
 
+
+        if ((requireActivity() as MainActivity).isWriter()) {
+            binding.fabAddItem.show()
+        }
+
+        binding.fabAddItem.setOnClickListener {
+
+            val alertDialog = AlertDialog.Builder(context).create()
+            val dialogAddItemBinding = DialogAddItemBinding.inflate(layoutInflater)
+            alertDialog.setView(dialogAddItemBinding.root)
+            alertDialog.setCancelable(true)
+            alertDialog.show()
+
+            dialogAddItemBinding.btnAdd.setOnClickListener {
+
+                if (dialogAddItemBinding.dialogAddEdtTitle.length() > 0 && dialogAddItemBinding.dialogEdtSubtitle.length() > 0 && dialogAddItemBinding.dialogAddEdtDetail.length() > 0 && dialogAddItemBinding.dialogAddEdtUrl.length() > 0) {
+                    val txtTitle = dialogAddItemBinding.dialogAddEdtTitle.text.toString()
+                    val txtSubtitle = dialogAddItemBinding.dialogEdtSubtitle.text.toString()
+                    val txtDetail = dialogAddItemBinding.dialogAddEdtDetail.text.toString()
+                    val txtUrl = dialogAddItemBinding.dialogAddEdtUrl.text.toString()
+                    val isTrend = dialogAddItemBinding.checkBoxTrend.isChecked
+                    val showExplore = dialogAddItemBinding.checkBoxExplore.isChecked
+                    val showGroup = dialogAddItemBinding.checkBoxGroups.isChecked
+                    val showOthers = dialogAddItemBinding.checkBoxOthers.isChecked
+
+                    val insight = if (isTrend) {
+                        val randomNum = (1..500).random()
+                        "+$randomNum K"
+                    } else {
+                        ""
+                    }
+                    alertDialog.dismiss()
+                    val item = ItemPost(txtUrl,txtTitle,txtSubtitle,txtDetail,isTrend,insight,showExplore,showGroup,showOthers)
+                    othersCloneList.add(0,item)
+                    (requireActivity() as MainActivity).getData().add(0,item)
+                    myAdapter.notifyItemInserted(0)
+                    binding.recyclerOthers.scrollToPosition(0)
+
+                } else {
+                    Toast.makeText(context, "Complete all parts", Toast.LENGTH_SHORT).show()
+                }
+            }
+
+
+        }
+
     }
 
     override fun onItemClicked(itemPost: ItemPost) {
@@ -78,25 +128,32 @@ class OthersFragment : Fragment(), ItemEvents {
 
     override fun onItemLongClicked(itemPost: ItemPost) {
         Toast.makeText(context, "${itemPost.txtTitle}", Toast.LENGTH_SHORT).show()
-        val sweetAlertDialog = SweetAlertDialog(context, SweetAlertDialog.WARNING_TYPE)
-        sweetAlertDialog.titleText = "Delete item"
-        sweetAlertDialog.confirmText = "Delete"
-        sweetAlertDialog.cancelText = "Cancel"
-        sweetAlertDialog.contentText = "want to delete this item?!"
 
-        sweetAlertDialog.setCancelClickListener {
-            sweetAlertDialog.dismiss()
+        if ((requireActivity() as MainActivity).isWriter()) {
+            val sweetAlertDialog = SweetAlertDialog(context, SweetAlertDialog.WARNING_TYPE)
+            sweetAlertDialog.titleText = "Delete item"
+            sweetAlertDialog.confirmText = "Delete"
+            sweetAlertDialog.cancelText = "Cancel"
+            sweetAlertDialog.contentText = "want to delete this item?!"
+
+            sweetAlertDialog.setCancelClickListener {
+                sweetAlertDialog.dismiss()
+            }
+
+            sweetAlertDialog.setConfirmClickListener {
+                (requireActivity() as MainActivity).deleteItem(itemPost)
+                myAdapter.notifyItemRemoved(othersCloneList.indexOf(itemPost))
+                othersCloneList.remove(itemPost)
+                sweetAlertDialog.dismiss()
+
+            }
+
+            sweetAlertDialog.show()
+        } else {
+
         }
 
-        sweetAlertDialog.setConfirmClickListener {
-            othersCloneList.remove(itemPost)
-            (requireActivity() as MainActivity).deleteItem(itemPost)
-            myAdapter.notifyItemRemoved(othersCloneList.indexOf(itemPost))
-            sweetAlertDialog.dismiss()
-
-        }
-
-        sweetAlertDialog.show()
     }
+
 
 }
