+++ app/src/main/java/com/example/breakingbadwiki/fragment/TrendFragment.kt
@@ -111,6 +111,51 @@ class TrendFragment : Fragment(), ItemEvents {
             LinearLayoutManager(context, RecyclerView.VERTICAL, false)
         binding.recyclerTrend.adapter = myAdapter
 
+        if ((requireActivity() as MainActivity).isWriter()) {
+            binding.fabAddItem.show()
+        }
+
+        binding.fabAddItem.setOnClickListener {
+
+            val alertDialog = AlertDialog.Builder(context).create()
+            val dialogAddItemBinding = DialogAddItemBinding.inflate(layoutInflater)
+            alertDialog.setView(dialogAddItemBinding.root)
+            alertDialog.setCancelable(true)
+            alertDialog.show()
+
+            dialogAddItemBinding.btnAdd.setOnClickListener {
+
+                if (dialogAddItemBinding.dialogAddEdtTitle.length() > 0 && dialogAddItemBinding.dialogEdtSubtitle.length() > 0 && dialogAddItemBinding.dialogAddEdtDetail.length() > 0 && dialogAddItemBinding.dialogAddEdtUrl.length() > 0) {
+                    val txtTitle = dialogAddItemBinding.dialogAddEdtTitle.text.toString()
+                    val txtSubtitle = dialogAddItemBinding.dialogEdtSubtitle.text.toString()
+                    val txtDetail = dialogAddItemBinding.dialogAddEdtDetail.text.toString()
+                    val txtUrl = dialogAddItemBinding.dialogAddEdtUrl.text.toString()
+                    val isTrend = dialogAddItemBinding.checkBoxTrend.isChecked
+                    val showExplore = dialogAddItemBinding.checkBoxExplore.isChecked
+                    val showGroup = dialogAddItemBinding.checkBoxGroups.isChecked
+                    val showOthers = dialogAddItemBinding.checkBoxOthers.isChecked
+
+                    val insight = if (isTrend) {
+                        val randomNum = (1..500).random()
+                        "+$randomNum K"
+                    } else {
+                        ""
+                    }
+                    alertDialog.dismiss()
+                    val item = ItemPost(txtUrl,txtTitle,txtSubtitle,txtDetail,isTrend,insight,showExplore,showGroup,showOthers)
+                    trendCloneList.add(0,item)
+                    (requireActivity() as MainActivity).getData().add(0,item)
+                    myAdapter.notifyItemInserted(0)
+                    binding.recyclerTrend.scrollToPosition(0)
+
+                } else {
+                    Toast.makeText(context, "Complete all parts", Toast.LENGTH_SHORT).show()
+                }
+            }
+
+
+        }
+
     }
 
     override fun onItemClicked(itemPost: ItemPost) {
@@ -121,25 +166,31 @@ class TrendFragment : Fragment(), ItemEvents {
 
     override fun onItemLongClicked(itemPost: ItemPost) {
         Toast.makeText(context, "${itemPost.txtTitle}", Toast.LENGTH_SHORT).show()
-        val sweetAlertDialog = SweetAlertDialog(context, SweetAlertDialog.WARNING_TYPE)
-        sweetAlertDialog.titleText = "Delete item"
-        sweetAlertDialog.confirmText = "Delete"
-        sweetAlertDialog.cancelText = "Cancel"
-        sweetAlertDialog.contentText = "want to delete this item?!"
 
-        sweetAlertDialog.setCancelClickListener {
-            sweetAlertDialog.dismiss()
+        if ((requireActivity() as MainActivity).isWriter() ) {
+            val sweetAlertDialog = SweetAlertDialog(context, SweetAlertDialog.WARNING_TYPE)
+            sweetAlertDialog.titleText = "Delete item"
+            sweetAlertDialog.confirmText = "Delete"
+            sweetAlertDialog.cancelText = "Cancel"
+            sweetAlertDialog.contentText = "want to delete this item?!"
+
+            sweetAlertDialog.setCancelClickListener {
+                sweetAlertDialog.dismiss()
+            }
+
+            sweetAlertDialog.setConfirmClickListener {
+                (requireActivity() as MainActivity).deleteItem(itemPost)
+                trendCloneList.remove(itemPost)
+                myAdapter.notifyItemRemoved(trendCloneList.indexOf(itemPost))
+                sweetAlertDialog.dismiss()
+
+            }
+
+            sweetAlertDialog.show()
+        } else {
+
         }
 
-        sweetAlertDialog.setConfirmClickListener {
-            (requireActivity() as MainActivity).deleteItem(itemPost)
-            trendCloneList.remove(itemPost)
-            myAdapter.notifyItemRemoved(trendCloneList.indexOf(itemPost))
-            sweetAlertDialog.dismiss()
-
-        }
-
-        sweetAlertDialog.show()
     }
 
 }
